Refactored my JSPs && added the EmailUtilClass && made changes to my controller by adding emailUtil

+++ src/main/java/com/bharath/location/repos/LocationRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.bharath.location.entities.Location;
 
-public interface LocationRepository extends JpaRepository<Location, Integer> {
+public interface LocationRepository extends JpaRepository<Location, Long> {
 
 }
